Extract out vpnd_client
